+++ packages/components/fields/async-select-field/src/async-select-field.visualroute.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Switch, Route } from 'react-router-dom';
 import { AsyncSelectField } from '@commercetools-frontend/ui-kit';
-import { Suite, Spec } from '../../../../test/percy';
+import { Suite, Spec } from '../../../../../test/percy';
 
 const loadOptions = () =>
   Promise.resolve([
