+++ packages/components/fields/async-select-field/src/index.js
@@ -1,0 +1,3 @@
+export { default } from './async-select-field';
+
+export { default as version } from './version';
