--- src/components/fields/async-select-field/index.js
@@ -1,1 +1,0 @@
-export { default } from './async-select-field';
